+++ src/renderer/src/components/onboarding/NotificationStep.test.tsx
@@ -1,7 +1,8 @@
 import { renderToStaticMarkup } from 'react-dom/server'
 import { describe, expect, it, vi } from 'vitest'
 import type { GlobalSettings } from '../../../../shared/types'
-import { NotificationStep, resolveMacNotificationPermissionState } from './NotificationStep'
+import { NotificationStep } from './NotificationStep'
+import { resolveMacNotificationPermissionState } from './mac-notification-permission-card'
 
 function createSettings(
   notificationOverrides: Partial<GlobalSettings['notifications']> = {}
