+++ src/renderer/src/components/onboarding/NotificationStep.tsx
@@ -1,7 +1,7 @@
-import { useCallback, useEffect, useRef, useState } from 'react'
-import { BellRing, Check, FileAudio, Settings, TriangleAlert, Upload } from 'lucide-react'
+import { useCallback, useRef, useState } from 'react'
+import { BellRing, FileAudio, Upload } from 'lucide-react'
 import { toast } from 'sonner'
-import type { GlobalSettings, NotificationDeliveryProbeResult } from '../../../../shared/types'
+import type { GlobalSettings } from '../../../../shared/types'
 import { Button } from '@/components/ui/button'
 import {
   Select,
@@ -13,6 +13,10 @@ import {
 } from '@/components/ui/select'
 import { sendNotificationSettingsTestNotification } from '@/components/settings/NotificationsPane'
 import { getNotificationSoundOptions } from '@/components/notification-sound-options'
+import {
+  MacNotificationPermissionCard,
+  useMacNotificationPermissionState
+} from '@/components/onboarding/mac-notification-permission-card'
 import { useMountedRef } from '@/hooks/useMountedRef'
 import { translate } from '@/i18n/i18n'
 
@@ -22,33 +26,6 @@ type NotificationStepProps = {
 }
 
 const CHOOSE_CUSTOM_SOUND_VALUE = 'choose-custom-file'
-
-export type MacNotificationPermissionState =
-  | 'checking'
-  | 'awaiting-permission'
-  | 'enabled'
-  | 'blocked'
-
-const MAC_PROBE_POLL_INTERVAL_MS = 2500
-// Why: bounded so an abandoned onboarding tab doesn't probe forever; ~3
-// minutes comfortably covers answering the dialog or flipping the toggle
-// in System Settings.
-const MAC_PROBE_POLL_MAX_ATTEMPTS = 72
-
-export function resolveMacNotificationPermissionState(
-  probeState: NotificationDeliveryProbeResult['state'],
-  promptedBefore: boolean
-): MacNotificationPermissionState | null {
-  if (probeState === 'unsupported') {
-    return null
-  }
-  if (probeState === 'delivered') {
-    return 'enabled'
-  }
-  // Why: a first-ever probe is what makes macOS show the permission dialog,
-  // so its rejection means "unanswered", not "denied".
-  return promptedBefore ? 'blocked' : 'awaiting-permission'
-}
 
 type NotificationSoundSelectValue =
   | GlobalSettings['notifications']['customSoundId']
@@ -66,8 +43,7 @@ export function NotificationStep({
 }: NotificationStepProps): React.JSX.Element {
   const notificationSettings = settings?.notifications
   const notificationSettingsRef = useRef(notificationSettings)
-  const [macPermissionState, setMacPermissionState] =
-    useState<MacNotificationPermissionState | null>(null)
+  const [macPermissionState, setMacPermissionState] = useMacNotificationPermissionState()
   const [isPickingSound, setIsPickingSound] = useState(false)
   const [selectPortalRoot, setSelectPortalRoot] = useState<HTMLElement | null>(null)
   const syncedNotificationSettingsRef = useRef(notificationSettings)
@@ -84,64 +60,6 @@ export function NotificationStep({
     // Why: onboarding sits above body-level portals, so the select menu must
     // portal into the overlay to stay clickable.
     setSelectPortalRoot(node?.closest<HTMLElement>('[data-onboarding-overlay]') ?? node)
-  }, [])
-
-  useEffect(() => {
-    let cancelled = false
-    let pollTimer: ReturnType<typeof setTimeout> | null = null
-    let pollAttempts = 0
-
-    // Why: probe rejections are silent, so polling while the card is not
-    // green costs nothing visible and lets the card flip to "enabled" the
-    // moment the user clicks Allow (or fixes the toggle in System Settings).
-    function schedulePoll(): void {
-      if (cancelled || pollAttempts >= MAC_PROBE_POLL_MAX_ATTEMPTS) {
-        return
-      }
-      pollTimer = setTimeout(() => {
-        pollAttempts += 1
-        void window.api.notifications.probeDelivery({ force: true }).then((probe) => {
-          if (cancelled) {
-            return
-          }
-          if (probe.state === 'delivered') {
-            setMacPermissionState('enabled')
-            return
-          }
-          schedulePoll()
-        })
-      }, MAC_PROBE_POLL_INTERVAL_MS)
-    }
-
-    void (async () => {
-      const status = await window.api.notifications.getPermissionStatus()
-      if (cancelled) {
-        return
-      }
-      if (status.platform !== 'darwin' || !status.supported) {
-        return
-      }
-      setMacPermissionState('checking')
-      // Why: `status.requested` is read before the probe stamps it, so a
-      // fresh install (where the probe itself pops the macOS dialog) renders
-      // as "answer the dialog" instead of "blocked".
-      const probe = await window.api.notifications.probeDelivery()
-      if (cancelled) {
-        return
-      }
-      const resolved = resolveMacNotificationPermissionState(probe.state, status.requested)
-      setMacPermissionState(resolved)
-      if (resolved === 'awaiting-permission' || resolved === 'blocked') {
-        schedulePoll()
-      }
-    })()
-
-    return () => {
-      cancelled = true
-      if (pollTimer) {
-        clearTimeout(pollTimer)
-      }
-    }
   }, [])
 
   const updateNotificationSettings = async (
@@ -255,103 +173,7 @@ export function NotificationStep({
 
   return (
     <div ref={setSelectPortalHost} className="space-y-5">
-      {macPermissionState === 'checking' ? (
-        <section className="rounded-xl border border-border bg-muted/20 px-5 py-4 text-[13px] text-muted-foreground">
-          {translate(
-            'auto.components.onboarding.NotificationStep.56b836215c',
-            'Checking notification permission…'
-          )}
-        </section>
-      ) : null}
-      {macPermissionState === 'enabled' ? (
-        <section className="flex items-center gap-2.5 rounded-xl border border-emerald-500/30 bg-emerald-500/[0.07] px-5 py-4">
-          <Check
-            className="size-4 shrink-0 text-emerald-600 dark:text-emerald-400"
-            strokeWidth={3}
-          />
-          <div className="min-w-0">
-            <div className="text-sm font-semibold text-foreground">
-              {translate(
-                'auto.components.onboarding.NotificationStep.fd84d3e9b8',
-                'Notifications are enabled'
-              )}
-            </div>
-            <p className="text-[13px] leading-relaxed text-muted-foreground">
-              {translate(
-                'auto.components.onboarding.NotificationStep.4f7bce5644',
-                'macOS will alert you when agents finish or terminals need attention.'
-              )}
-            </p>
-          </div>
-        </section>
-      ) : null}
-      {macPermissionState === 'awaiting-permission' ? (
-        <section className="rounded-xl border border-border bg-card px-5 py-4">
-          <div className="flex flex-wrap items-start justify-between gap-4">
-            <div className="min-w-0 space-y-1">
-              <div className="flex items-center gap-2 text-sm font-semibold text-foreground">
-                <BellRing className="size-4" />
-                {translate(
-                  'auto.components.onboarding.NotificationStep.95d99b52fa',
-                  'Allow notifications for Orca'
-                )}
-              </div>
-              <p className="max-w-[58ch] text-[13px] leading-relaxed text-muted-foreground">
-                {translate(
-                  'auto.components.onboarding.NotificationStep.94562ba367',
-                  'macOS is asking for permission. Click Allow in the dialog and this step updates automatically.'
-                )}
-              </p>
-            </div>
-            <Button
-              type="button"
-              variant="outline"
-              size="sm"
-              className="gap-2"
-              onClick={() => void window.api.notifications.openSystemSettings()}
-            >
-              <Settings className="size-3.5" />
-              {translate(
-                'auto.components.onboarding.NotificationStep.4f6a1da718',
-                'Open System Settings'
-              )}
-            </Button>
-          </div>
-        </section>
-      ) : null}
-      {macPermissionState === 'blocked' ? (
-        <section className="rounded-xl border border-amber-500/30 bg-amber-500/[0.07] px-5 py-4">
-          <div className="flex flex-wrap items-start justify-between gap-4">
-            <div className="min-w-0 space-y-1">
-              <div className="flex items-center gap-2 text-sm font-semibold text-foreground">
-                <TriangleAlert className="size-4 text-amber-600 dark:text-amber-400" />
-                {translate(
-                  'auto.components.onboarding.NotificationStep.90b5d2e363',
-                  'macOS is not delivering Orca notifications'
-                )}
-              </div>
-              <p className="max-w-[58ch] text-[13px] leading-relaxed text-muted-foreground">
-                {translate(
-                  'auto.components.onboarding.NotificationStep.2c47f5465f',
-                  'Turn on Allow notifications for Orca in System Settings. This step updates automatically once enabled.'
-                )}
-              </p>
-            </div>
-            <Button
-              type="button"
-              size="sm"
-              className="gap-2"
-              onClick={() => void window.api.notifications.openSystemSettings()}
-            >
-              <Settings className="size-3.5" />
-              {translate(
-                'auto.components.onboarding.NotificationStep.4f6a1da718',
-                'Open System Settings'
-              )}
-            </Button>
-          </div>
-        </section>
-      ) : null}
+      <MacNotificationPermissionCard state={macPermissionState} />
 
       <section className="space-y-3">
         <div className="space-y-1">
